patch228

See merge request oceanbase-ce-publish/obconnector-c!2

+++ libmariadb/ob_utils.c
@@ -32,5 +32,5 @@ int64_t get_current_time_us()
 {
   struct timeval tv;
   gettimeofday(&tv, NULL);
-  return tv.tv_sec * 1000000 + tv.tv_usec;
+  return (int64_t)tv.tv_sec * 1000000 + tv.tv_usec;
 }
